rename hussar to server

+++ src/main.cpp
@@ -6,7 +6,7 @@
 int main(int argc, char* argv[])
 {
     const std::string host = "127.0.0.1";
-    const unsigned int port = 8080;
+    const short port = 8080;
 
     Hussar hussar(host, port);
